requirements.txt
certifi==2020.12.5
chardet==4.0.0
facebook-sdk==3.1.0
idna==2.10
PyMySQL==1.0.2
requests==2.25.1
SQLAlchemy==1.3.23
urllib3==1.26.3
